refactor: make importing templates available as a lib
