+++ govc/flags/output.go
@@ -1,11 +1,11 @@
 /*
-Copyright (c) 2014-2016 VMware, Inc. All Rights Reserved.
+Copyright (c) 2014-2024 VMware, Inc. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
@@ -26,7 +26,6 @@ import (
 	"os"
 	"reflect"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/dougm/pretty"
@@ -204,7 +203,7 @@ type errorOutput struct {
 }
 
 func (e errorOutput) Write(w io.Writer) error {
-	reason := e.error.Error()
+	reason := e.Error()
 	var messages []string
 	var faults []types.LocalizableMessage
 
@@ -261,15 +260,15 @@ func (e errorOutput) canEncode() bool {
 	return soap.IsSoapFault(e.error) || soap.IsVimFault(e.error)
 }
 
-// cannotEncode causes cli.Run to output err.Error() as it would without an error format specified
-var cannotEncode = errors.New("cannot encode error")
+// errCannotEncode causes cli.Run to output err.Error() as it would without an error format specified
+var errCannotEncode = errors.New("cannot encode error")
 
 func (e errorOutput) MarshalJSON() ([]byte, error) {
 	_, ok := e.error.(json.Marshaler)
 	if ok || e.canEncode() {
 		return json.Marshal(e.error)
 	}
-	return nil, cannotEncode
+	return nil, errCannotEncode
 }
 
 func (e errorOutput) MarshalXML(encoder *xml.Encoder, start xml.StartElement) error {
@@ -277,108 +276,9 @@ func (e errorOutput) MarshalXML(encoder *xml.Encoder, start xml.StartElement) er
 	if ok || e.canEncode() {
 		return encoder.Encode(e.error)
 	}
-	return cannotEncode
+	return errCannotEncode
 }
 
-type progressLogger struct {
-	flag   *OutputFlag
-	prefix string
-
-	wg sync.WaitGroup
-
-	sink chan chan progress.Report
-	done chan struct{}
-}
-
-func newProgressLogger(flag *OutputFlag, prefix string) *progressLogger {
-	p := &progressLogger{
-		flag:   flag,
-		prefix: prefix,
-
-		sink: make(chan chan progress.Report),
-		done: make(chan struct{}),
-	}
-
-	p.wg.Add(1)
-
-	go p.loopA()
-
-	return p
-}
-
-// loopA runs before Sink() has been called.
-func (p *progressLogger) loopA() {
-	var err error
-
-	defer p.wg.Done()
-
-	tick := time.NewTicker(100 * time.Millisecond)
-	defer tick.Stop()
-
-	called := false
-
-	for stop := false; !stop; {
-		select {
-		case ch := <-p.sink:
-			err = p.loopB(tick, ch)
-			stop = true
-			called = true
-		case <-p.done:
-			stop = true
-		case <-tick.C:
-			line := fmt.Sprintf("\r%s", p.prefix)
-			p.flag.Log(line)
-		}
-	}
-
-	if err != nil && err != io.EOF {
-		p.flag.Log(fmt.Sprintf("\r%sError: %s\n", p.prefix, err))
-	} else if called {
-		p.flag.Log(fmt.Sprintf("\r%sOK\n", p.prefix))
-	}
-}
-
-// loopA runs after Sink() has been called.
-func (p *progressLogger) loopB(tick *time.Ticker, ch <-chan progress.Report) error {
-	var r progress.Report
-	var ok bool
-	var err error
-
-	for ok = true; ok; {
-		select {
-		case r, ok = <-ch:
-			if !ok {
-				break
-			}
-			err = r.Error()
-		case <-tick.C:
-			line := fmt.Sprintf("\r%s", p.prefix)
-			if r != nil {
-				line += fmt.Sprintf("(%.0f%%", r.Percentage())
-				detail := r.Detail()
-				if detail != "" {
-					line += fmt.Sprintf(", %s", detail)
-				}
-				line += ")"
-			}
-			p.flag.Log(line)
-		}
-	}
-
-	return err
-}
-
-func (p *progressLogger) Sink() chan<- progress.Report {
-	ch := make(chan progress.Report)
-	p.sink <- ch
-	return ch
-}
-
-func (p *progressLogger) Wait() {
-	close(p.done)
-	p.wg.Wait()
-}
-
-func (flag *OutputFlag) ProgressLogger(prefix string) *progressLogger {
-	return newProgressLogger(flag, prefix)
+func (flag *OutputFlag) ProgressLogger(prefix string) *progress.ProgressLogger {
+	return progress.NewProgressLogger(flag.Log, prefix)
 }
