+++ govc/importx/ova.go
@@ -1,11 +1,11 @@
 /*
-Copyright (c) 2014-2015 VMware, Inc. All Rights Reserved.
+Copyright (c) 2014-2024 VMware, Inc. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
@@ -22,6 +22,7 @@ import (
 
 	"github.com/vmware/govmomi/govc/cli"
 	"github.com/vmware/govmomi/object"
+	"github.com/vmware/govmomi/ovf/importer"
 	"github.com/vmware/govmomi/vim25/types"
 )
 
@@ -43,21 +44,21 @@ func (cmd *ova) Run(ctx context.Context, f *flag.FlagSet) error {
 		return err
 	}
 
-	archive := &TapeArchive{Path: fpath}
-	archive.Client = cmd.Client
+	archive := &importer.TapeArchive{Path: fpath}
+	archive.Client = cmd.Importer.Client
 
-	cmd.Archive = archive
+	cmd.Importer.Archive = archive
 
 	moref, err := cmd.Import(fpath)
 	if err != nil {
 		return err
 	}
 
-	vm := object.NewVirtualMachine(cmd.Client, *moref)
+	vm := object.NewVirtualMachine(cmd.Importer.Client, *moref)
 	return cmd.Deploy(vm, cmd.OutputFlag)
 }
 
 func (cmd *ova) Import(fpath string) (*types.ManagedObjectReference, error) {
 	ovf := "*.ovf"
-	return cmd.ovfx.Import(ovf)
+	return cmd.Importer.Import(context.TODO(), ovf, cmd.Options)
 }
