+++ ovf/importer/archive.go
@@ -1,11 +1,11 @@
 /*
-Copyright (c) 2014-2015 VMware, Inc. All Rights Reserved.
+Copyright (c) 2024-2024 VMware, Inc. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,14 +14,13 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package importx
+package importer
 
 import (
 	"archive/tar"
 	"bytes"
 	"context"
 	"errors"
-	"flag"
 	"fmt"
 	"io"
 	"net/url"
@@ -31,32 +30,12 @@ import (
 	"strings"
 
 	"github.com/vmware/govmomi/ovf"
-	"github.com/vmware/govmomi/vapi/library"
 	"github.com/vmware/govmomi/vim25"
 	"github.com/vmware/govmomi/vim25/soap"
 )
 
-// ArchiveFlag doesn't register any flags;
-// only encapsulates some common archive related functionality.
-type ArchiveFlag struct {
-	Archive
-
-	manifest map[string]*library.Checksum
-}
-
-func newArchiveFlag(ctx context.Context) (*ArchiveFlag, context.Context) {
-	return &ArchiveFlag{}, ctx
-}
-
-func (f *ArchiveFlag) Register(ctx context.Context, fs *flag.FlagSet) {
-}
-
-func (f *ArchiveFlag) Process(ctx context.Context) error {
-	return nil
-}
-
-func (f *ArchiveFlag) ReadOvf(fpath string) ([]byte, error) {
-	r, _, err := f.Open(fpath)
+func ReadOvf(fpath string, a Archive) ([]byte, error) {
+	r, _, err := a.Open(fpath)
 	if err != nil {
 		return nil, err
 	}
@@ -65,29 +44,13 @@ func (f *ArchiveFlag) ReadOvf(fpath string) ([]byte, error) {
 	return io.ReadAll(r)
 }
 
-func (f *ArchiveFlag) ReadEnvelope(data []byte) (*ovf.Envelope, error) {
+func ReadEnvelope(data []byte) (*ovf.Envelope, error) {
 	e, err := ovf.Unmarshal(bytes.NewReader(data))
 	if err != nil {
 		return nil, fmt.Errorf("failed to parse ovf: %s", err)
 	}
 
 	return e, nil
-}
-
-func (f *ArchiveFlag) readManifest(fpath string) error {
-	base := filepath.Base(fpath)
-	ext := filepath.Ext(base)
-	mfName := strings.Replace(base, ext, ".mf", 1)
-
-	mf, _, err := f.Open(mfName)
-	if err != nil {
-		msg := fmt.Sprintf("manifest %q: %s", mf, err)
-		fmt.Fprintln(os.Stderr, msg)
-		return errors.New(msg)
-	}
-	f.manifest, err = library.ReadManifest(mf)
-	_ = mf.Close()
-	return err
 }
 
 type Archive interface {
@@ -163,7 +126,7 @@ type Opener struct {
 	*vim25.Client
 }
 
-func isRemotePath(path string) bool {
+func IsRemotePath(path string) bool {
 	if strings.HasPrefix(path, "http://") || strings.HasPrefix(path, "https://") {
 		return true
 	}
@@ -185,7 +148,7 @@ func (o Opener) OpenLocal(path string) (io.ReadCloser, int64, error) {
 }
 
 func (o Opener) OpenFile(path string) (io.ReadCloser, int64, error) {
-	if isRemotePath(path) {
+	if IsRemotePath(path) {
 		return o.OpenRemote(path)
 	}
 	return o.OpenLocal(path)
